requirements.txt
Django==4.2
celery==5.2.7
redis==4.5.5
requests==2.31.0
djangorestframework==3.14.0


